dakota+thundr: Fix pixel positioning of text
This limits viewports to pixel boundaries. This prevents scenarios
where something that expects to present itself at normal pixel boundaries
gets horribly aliased from being offset by 0.018. This is a pretty
normal approach for text rendering.

Tested with DT. Fixes text

+++ src/category5/vkcomp/wm/mod.rs
@@ -607,7 +607,7 @@ impl WindowManager {
     fn begin_frame(&mut self, atmos: &mut Atmosphere) -> Result<()> {
         self.record_draw(atmos);
         let res = self.wm_thundr.get_resolution();
-        let viewport = th::Viewport::new(0.0, 0.0, res.0 as f32, res.1 as f32);
+        let viewport = th::Viewport::new(0, 0, res.0 as i32, res.1 as i32);
 
         // First push our CPU data to the GPU
         self.wm_thundr.flush_surface_data(&mut self.wm_surfaces)?;
